Add files via upload
Cleaned up messy code and added some good practices. Code is now properly object-oriented but some bugs remain. 
This new version includes changes in the analysis workflow, namely the program now outputs the raw and treated I-V STS curves instead of only the derivatives
Bugs that remain to be treated in the near future include treatment of edge cases in the method for discretizing and selecting blocks on the topography image overlay, as well as rehashing redundancies in the map generator method.
Next orders of business shall be the implementation of Reinforcement Machine learning for automated clustering of peaks in the di/dV curves as well as features on the topography, suggesting the user take a closer look at specific correlations